Перенёс общие функции в отдельный документ
