--- src/utils.js
@@ -1,45 +1,0 @@
-import dayjs from 'dayjs';
-export const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-export const getRandomItem = (items) => {
-  if (items.length === 0) {
-    return [];
-  }
-  return items[getRandomInteger(0, items.length - 1)];
-};
-
-export const getFormatedDate = (date, format) => {
-  return dayjs(date).format(format);
-};
-
-// --- Рендер ---
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`); // 1
-  newElement.innerHTML = template; // 2
-
-  return newElement.firstChild; // 3
-};
-
-
